+++ server/s_synth.hpp
@@ -10,21 +10,21 @@
 class Synth : public Source, public Entity
 {
 public:
-  Synth(const float &bpm, class Sink &, OscType, Envelope);
+  Synth(const double &bpm, class Sink &, OscType, Envelope);
   auto pull(int samples) -> std::vector<float>;
   auto operator()(Note) -> void;
   auto operator()(Envelope) -> void;
   auto isBusy() const -> bool final;
 
 private:
-  std::reference_wrapper<const float> bpm;
+  std::reference_wrapper<const double> bpm;
   OscType oscType;
   Envelope envelope;
   int pos = 0;
   struct N
   {
-    float note;
-    float vel = 1.f;
+    double note;
+    double vel = 1.;
     int start;
     int end;
   };
