+++ shared/envelope.hpp
@@ -3,10 +3,10 @@
 
 struct Envelope
 {
-  float attack = 0.01;                                   // sec
-  float decay = 0.1;                                     // sec
-  float sustain = -20;                                   // dB
-  float release = 0.1;                                   // sec
-  auto amp(float sample, float duration) const -> float; // 0 - 1
+  double attack = 0.01;                                     // sec
+  double decay = 0.1;                                       // sec
+  double sustain = -20;                                     // dB
+  double release = 0.1;                                     // sec
+  auto amp(double sample, double duration) const -> double; // 0 - 1
   SER_PROPS(attack, decay, sustain, release);
 };
